+++ src/main/java/com/trs/gov/kpi/entity/responsedata/WkOneSiteScoreResponse.java
@@ -12,7 +12,7 @@ public class WkOneSiteScoreResponse {
 
     private Integer checkId;//检查编号
 
-    private Integer total;//综合评分
+    private double total;//综合评分
 
     private Integer invalidLink;//链接可用性分数
